Sync with delta

+++ showcase/src/example/models/models_yaw_pitch_roll.rs
@@ -160,7 +160,7 @@ pub fn run(rl: &mut RaylibHandle, thread: &RaylibThread) -> crate::SampleOut {
                         0.0,
                         0.0,
                         tex_pitch.width() as f32,
-                        tex_pitch.height() as f32 as f32,
+                        tex_pitch.height() as f32,
                     ),
                     Rectangle::new(
                         center_x,
@@ -183,7 +183,7 @@ pub fn run(rl: &mut RaylibHandle, thread: &RaylibThread) -> crate::SampleOut {
                         0.0,
                         0.0,
                         tex_plane.width() as f32,
-                        tex_plane.height() as f32 as f32,
+                        tex_plane.height() as f32,
                     ),
                     Rectangle::new(
                         center_x,
